refactor(plugins): extract IPC handlers and remove debug logs
- Extract 8 plugin IPC handlers from main.ts into ipcHandlers/plugins/
- Remove 14 verbose debug console.log statements from pluginManager,
  PluginsSettings, and main.ts (retain warnings and operational logs)
- Run eslint --fix on all PR-modified files
