+++ src/main/ipcHandlers/plugins/index.ts
@@ -1,0 +1,2 @@
+export type { PluginHandlerDeps } from './handlers';
+export { registerPluginHandlers } from './handlers';
